+++ desktop/cmp/button/AppMenuButton.js
@@ -4,12 +4,13 @@
  *
  * Copyright © 2020 Extremely Heavy Industries Inc.
  */
-
+import React from 'react';
 import PT from 'prop-types';
 import {hoistCmp, XH} from '@xh/hoist/core';
 import {menu, menuItem, menuDivider, popover} from '@xh/hoist/kit/blueprint';
 import {Button, button} from '@xh/hoist/desktop/cmp/button';
 import {Icon} from '@xh/hoist/icon';
+import {filterConsecutiveMenuSeparators} from '@xh/hoist/utils/impl';
 
 export const [AppMenuButton, appMenuButton] = hoistCmp.withFactory({
     displayName: 'AppMenuButton',
@@ -17,17 +18,7 @@ export const [AppMenuButton, appMenuButton] = hoistCmp.withFactory({
     className: 'xh-app-menu',
 
     render(props) {
-        let {className, hideOptionsItem, hideFeedbackItem, hideThemeItem, hideAdminItem, hideImpersonateItem, hideLogoutItem, extraItems, ...rest} = props;
-        extraItems = extraItems ?
-            [...extraItems.map(m => menuItem(m)), menuDivider()] :
-            [];
-
-        hideAdminItem = hideAdminItem || !XH.getUser().isHoistAdmin;
-        hideImpersonateItem = hideImpersonateItem || !XH.identityService.canImpersonate;
-        hideLogoutItem = hideLogoutItem || XH.appSpec.isSSO;
-        hideOptionsItem = hideOptionsItem || !XH.acm.optionsDialogModel.hasOptions;
-
-        // TODO:  Need logic from context menu to remove duplicate separators!
+        const {className, ...rest} = props;
         return popover({
             className,
             position: 'bottom-right',
@@ -36,55 +27,15 @@ export const [AppMenuButton, appMenuButton] = hoistCmp.withFactory({
                 icon: Icon.bars(),
                 ...rest
             }),
-            content: menu(
-                ...extraItems,
-                menuItem({
-                    omit: hideOptionsItem,
-                    text: 'Options',
-                    icon: Icon.options(),
-                    onClick: () => XH.showOptionsDialog()
-                }),
-                menuItem({
-                    omit: hideFeedbackItem,
-                    text: 'Feedback',
-                    icon: Icon.comment({className: 'fa-flip-horizontal'}),
-                    onClick: () => XH.showFeedbackDialog()
-                }),
-                menuItem({
-                    omit: hideThemeItem,
-                    text: XH.darkTheme ? 'Light Theme' : 'Dark Theme',
-                    icon: XH.darkTheme ? Icon.sun({prefix: 'fas'}) : Icon.moon(),
-                    onClick: () => XH.toggleTheme()
-                }),
-                menuDivider({omit: hideAdminItem && hideImpersonateItem}),
-                menuItem({
-                    omit: hideAdminItem,
-                    text: 'Admin',
-                    icon: Icon.wrench(),
-                    onClick: () => window.open('/admin')
-                }),
-                menuItem({
-                    omit: hideImpersonateItem,
-                    text: 'Impersonate',
-                    icon: Icon.impersonate(),
-                    onClick: () => XH.showImpersonationBar()
-                }),
-                menuDivider({omit: hideLogoutItem}),
-                menuItem({
-                    omit: hideLogoutItem,
-                    text: 'Logout',
-                    icon: Icon.logout(),
-                    intent: 'danger',
-                    onClick: () => XH.identityService.logoutAsync()
-                })
-            )
+            content: menu(buildMenuItems(props))
         });
     }
 });
+
 AppMenuButton.propTypes = {
     ...Button.propTypes,
 
-    /** Array of configs for any app-specific menu items. */
+    /** Array of configs or valid React Elements (such as menuDividers) for any app-specific menu items. */
     extraItems: PT.array,
 
     /** True to hide the Admin Item. Always hidden for users w/o HOIST_ADMIN role. */
@@ -108,3 +59,75 @@ AppMenuButton.propTypes = {
     /** True to hide the Theme Toggle button. */
     hideThemeItem: PT.bool
 };
+
+//---------------------------
+// Implementation
+//---------------------------
+function buildMenuItems({
+    hideOptionsItem,
+    hideFeedbackItem,
+    hideThemeItem,
+    hideAdminItem,
+    hideImpersonateItem,
+    hideLogoutItem,
+    extraItems = []
+}) {
+    hideOptionsItem = hideOptionsItem || !XH.acm.optionsDialogModel.hasOptions;
+    hideAdminItem = hideAdminItem || !XH.getUser().isHoistAdmin;
+    hideImpersonateItem = hideImpersonateItem || !XH.identityService.canImpersonate;
+    hideLogoutItem = hideLogoutItem || XH.appSpec.isSSO;
+
+    const defaultItems = [
+        {
+            omit: hideOptionsItem,
+            text: 'Options',
+            icon: Icon.options(),
+            onClick: () => XH.showOptionsDialog()
+        },
+        {
+            omit: hideFeedbackItem,
+            text: 'Feedback',
+            icon: Icon.comment({className: 'fa-flip-horizontal'}),
+            onClick: () => XH.showFeedbackDialog()
+        },
+        {
+            omit: hideThemeItem,
+            text: XH.darkTheme ? 'Light Theme' : 'Dark Theme',
+            icon: XH.darkTheme ? Icon.sun({prefix: 'fas'}) : Icon.moon(),
+            onClick: () => XH.toggleTheme()
+        },
+        '-',
+        {
+            omit: hideAdminItem,
+            text: 'Admin',
+            icon: Icon.wrench(),
+            onClick: () => window.open('/admin')
+        },
+        {
+            omit: hideImpersonateItem,
+            text: 'Impersonate',
+            icon: Icon.impersonate(),
+            onClick: () => XH.showImpersonationBar()
+        },
+        '-',
+        {
+            omit: hideLogoutItem,
+            text: 'Logout',
+            icon: Icon.logout(),
+            intent: 'danger',
+            onClick: () => XH.identityService.logoutAsync()
+        }
+    ];
+
+    return [
+        ...extraItems,
+        '-',
+        ...defaultItems
+    ]
+        .filter(it => !it.omit)
+        .filter(filterConsecutiveMenuSeparators())
+        .map(it => {
+            if (it === '-') return menuDivider();
+            return React.isValidElement(it) ? it : menuItem(it);
+        });
+}